pyproject.toml
[project]
name = "llm-project"
version = "0.1.0"
description = "Boot.dev course: Build an AI agent in Python"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai==1.12.1",
    "python-dotenv==1.1.0",
]
